[Router] Restore search results skeleton, refactor
