+++ src/desktop/assets/experimental-app-shell.tsx
@@ -1,2 +1,1 @@
-// FIXME: Remove -- this can go in global bundle
 import "desktop/apps/experimental-app-shell/client"
