+++ src/desktop/components/main_layout/templates/blank.jade
@@ -21,11 +21,16 @@ html( data-useragent= userAgent lang="en")
     if options.flash
       include ../../flash/template
 
+    //- Needed for search results skeleton
+    if pageType == 'search'
+      #search-page-header
+        include ../header/templates/index
+
     #blank-layout-container: block body
 
     #scripts
-      include scripts
-      script( src= asset('/assets/main_layout.js') )
-      if assetPackage
-        script( src= asset('/assets/#{assetPackage}.js') )
       block scripts
+        include scripts
+        script( src= asset('/assets/main_layout.js') )
+        if assetPackage
+          script( src= asset('/assets/#{assetPackage}.js') )
