+++ src/desktop/components/main_layout/templates/react_blank_index.jade
@@ -12,6 +12,9 @@ block head
 block body
   #react-root
     != body
+  if pageType == 'search'
+    #loading-container
+      != loadingComponent
   #react-portal
 
   if jsonLD
